// src/app/models/contract.model.ts
export interface Contract {
  id: String;
  title: String;
  business: String;
  process: String;
  numberOfMachines: String;
  line: String;
  date: String;
  asset: String;
  system: String;
  assetOrCriticComponent: String;
  inventory: String;
  systemFunction: String;
  failtureMode: String;
  faultureOrigin: String;
  operationalCheckList: String;
  preventivePlans: String;
  predictivePlans: String;
  category: String;
  workPlace: String;
  frecuency: String;
  lineCondition: String;
  hm: String;
  description: String;
}